+++ .github/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## ✨ 1.7.6 - Adds Multi-Search Support with Bangs [PR #224](https://github.com/Lissy93/dashy/pull/224)
+- Adds option for user to add custom search bangs, in order to specify search engine/ target app. Re: #206
+
 ## 🎨 1.7.5 - Improved Language Detection & UI [PR #223](https://github.com/Lissy93/dashy/pull/223)
 - Makes the auto language detection algo smarter
 - Improves responsiveness for the language selector form